refactor(engine): break up building
